src/server/pyproject.toml
[tool.poetry]
name = "langstormapp"
version = "0.1.0"
description = ""
authors = ["Vedernikov Andrey <vedernikov.ai.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
